@@ -42,11 +42,12 @@
     "@react-three/fiber": "^7.0.26",
     "chroma-js": "^3.0.0",
     "d3-array": "^3.1.1",
+    "d3-scale": "^4.0.2",
     "d3-selection": "^3.0.0",
     "d3-transition": "^3.0.1",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "three": "^0.137.5"
   }
 }
