@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import Site from "./Site.js";
 import './assets/css/style.css';
 import { createTheme, ThemeProvider } from '@mui/material';
@@ -15,9 +15,13 @@ const theme = createTheme({
   }
 });
 
-ReactDOM.render(
-  <ThemeProvider theme={theme}>
-    <Site />
-  </ThemeProvider>,
-  document.getElementById('root')
+// Use createRoot to handle the root element
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <ThemeProvider theme={theme}>
+      <Site />
+    </ThemeProvider>
+  </React.StrictMode>
 );
+
